# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardStats {
  totalOrders: number;
  totalAuthorizedUsers: number;
  totalTouches: number;
  activeAccounts: number;
  totalCreditsHeld: number;
  totalRevenueCents: number;
}

export interface Order {
  id: number;
  discordId: string;
  orderUuid: string;
  ddAccountId: number;
  /** @nullable */
  groupOrderUrl?: string | null;
  /** @nullable */
  storeName?: string | null;
  totalCents: number;
  creditsUsed: number;
  status: string;
  /** @nullable */
  trackingUrl?: string | null;
  createdAt: string;
}

export interface DdAccount {
  id: number;
  name: string;
  email: string;
  isActive: boolean;
  addedBy: string;
  createdAt: string;
}

export interface ToggleResult {
  ok: boolean;
  isActive: boolean;
}

export interface AuthorizedUser {
  id: number;
  discordId: string;
  discordUsername: string;
  addedBy: string;
  addedAt: string;
}

export interface Touch {
  id: number;
  discordId: string;
  discordUsername: string;
  photoUrl: string;
  messageId: string;
  channelId: string;
  guildId: string;
  /** @nullable */
  note?: string | null;
  createdAt: string;
}

export interface UserCredit {
  id: number;
  discordId: string;
  balance: number;
  updatedAt: string;
}

export interface BotConfig {
  id: number;
  ownerDiscordId: string;
  createdAt: string;
}

export type GetDashboardOrdersParams = {
limit?: number;
offset?: number;
};