# lib/api-zod/src/generated/types/authorizedUser.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuthorizedUser {
  id: number;
  discordId: string;
  discordUsername: string;
  addedBy: string;
  addedAt: string;
}