# lib/api-zod/src/generated/types/botConfig.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BotConfig {
  id: number;
  ownerDiscordId: string;
  createdAt: string;
}