# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  totalOrders: number;
  totalAuthorizedUsers: number;
  totalTouches: number;
  activeAccounts: number;
  totalCreditsHeld: number;
  totalRevenueCents: number;
}