# lib/api-zod/src/generated/types/ddAccount.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DdAccount {
  id: number;
  name: string;
  email: string;
  isActive: boolean;
  addedBy: string;
  createdAt: string;
}