# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Order {
  id: number;
  discordId: string;
  orderUuid: string;
  ddAccountId: number;
  /** @nullable */
  groupOrderUrl?: string | null;
  /** @nullable */
  storeName?: string | null;
  totalCents: number;
  creditsUsed: number;
  status: string;
  /** @nullable */
  trackingUrl?: string | null;
  createdAt: string;
}