# lib/api-zod/src/generated/types/touch.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Touch {
  id: number;
  discordId: string;
  discordUsername: string;
  photoUrl: string;
  messageId: string;
  channelId: string;
  guildId: string;
  /** @nullable */
  note?: string | null;
  createdAt: string;
}