# lib/api-zod/src/generated/types/userCredit.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserCredit {
  id: number;
  discordId: string;
  balance: number;
  updatedAt: string;
}